spawn_essential on SpawnHandle is now an Ext auto-impl trait
