+++ clique/src/lib.rs
@@ -24,7 +24,7 @@ mod testing;
 
 pub use crypto::{PublicKey, SecretKey};
 pub use rate_limiting::{RateLimitingDialer, RateLimitingListener};
-pub use service::{Service, SpawnHandleT};
+pub use service::{Service, SpawnHandleExt, SpawnHandleT};
 
 const LOG_TARGET: &str = "network-clique";
 /// A basic alias for properties we expect basic data to satisfy.
