+++ finality-aleph/src/abft/current/mod.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 use current_aleph_bft::{create_config, default_delay_config, Config, LocalIO, Terminator};
 use log::debug;
-use network_clique::SpawnHandleT;
+use network_clique::SpawnHandleExt;
 
 mod network;
 mod traits;
