+++ finality-aleph/src/party/manager/task.rs
@@ -2,7 +2,7 @@ use std::{boxed::Box, pin::Pin};
 
 use futures::channel::oneshot;
 use log::{debug, warn};
-use network_clique::SpawnHandleT;
+use network_clique::SpawnHandleExt;
 
 use crate::{Future, SpawnHandle};
